refactor EmptyTrashButton
